+++ web/package.json
@@ -11,9 +11,9 @@
     "clear": "docusaurus clear",
     "deploy": "docusaurus deploy",
     "docusaurus": "docusaurus",
-    "generate-md": "npm run generate-md-docs && npm run generate-md-llm-files",
-    "generate-md-docs": "tsx ./scripts/markdown-docs/html-to-md/index.ts",
-    "generate-md-llm-files": "tsx ./scripts/markdown-docs/llm-files/generate-llm-files.ts",
+    "generate-md": "npm run generate-md:docs && npm run generate-md:llm-files",
+    "generate-md:docs": "tsx ./scripts/markdown-docs/html-to-md/index.ts",
+    "generate-md:llm-files": "tsx ./scripts/markdown-docs/generate-llm-files/index.ts",
     "install-spec-package": "npm install --prefix ../waspc/data/packages/spec",
     "remark": "remark",
     "remark:once": "npm run remark -- --quiet --frail --use remark-validate-links docs/",
