+++ web/scripts/markdown-docs/generate-llm-files/constants.ts
@@ -1,0 +1,2 @@
+// TODO: Maybe localhost for local development? Easier to test.
+export const WASP_BASE_URL = "https://wasp.sh";
