--- web/scripts/markdown-docs/llm-files/constants.ts
@@ -1,2 +1,0 @@
-// TODO: maybe localhost for local development?
-export const WASP_BASE_URL = "https://wasp.sh";
